Support for slicing and index access on DataSources and Layers: for layer in ds:    do something
for feature in layer:
   do something

This also means that we get list comprehension support
[i.GetName() for i in ds[0:4]]
or
[i..GetFID() for feature in layer[0:]]

*Note*-- some datasources do not always have sequential
feature ids and slicing may not work as expected for all
datasources.

Support for len(ds) or len(layer) to return the number of
layers in the datasource or features in the layer (these are
simple wrappers around GetLayerCount() and GetFeatureCount())

A custom exception, OGRError, is now raised in many cases
instead of generic IndexErrors or ValueErrors when the error
is really coming from OGR not being able to read something.

Reworked DataSource.GetLayer() to check the types of the input
value (int or string) and call the appropriate OGR function to do so.
__getitem__ in DataSource uses this function instead of switching
between it and GetLayerByName().


git-svn-id: https://svn.osgeo.org/gdal@6980 f0d54148-0727-0410-94bb-9a71ac55c965