Add a check that the CreateDataSource result is not none.
Fixes pyflakes complaint for a driver CreateDataSource call that is immediately followed by a
deletion request of the ds (by setting it to None).  r29697 changed "del ds" -> "ds = None"
to match the convention dominant in the rest of the ogr test directory.  It does not hurt to
add a check that immediately verifies that the CreateDataSource call succeeded.

Example error:

pyflakes autotest
autotest/ogr/ogr_libkml.py:1070: local variable 'ds' is assigned to but never used


git-svn-id: https://svn.osgeo.org/gdal@29699 f0d54148-0727-0410-94bb-9a71ac55c965