static int foo() -> static bool fool().
CPL_RectContained(), CPL_RectOverlap(), and CPLQuadTreeNodeForeach().

Switch static int functions that are really bools to bools.  Because
these functions are internal only, this change does not alter the C or
C++ ABIs.  The change makes it easier to see that the functions really
only return true/false.  This will help static analyzers by more
rapidly constraining the execution state space.  Additionally, it's
possible that some compilers may be be able to do a better job
optimizing the resulting object code.


git-svn-id: https://svn.osgeo.org/gdal@30816 f0d54148-0727-0410-94bb-9a71ac55c965